requirements.txt
gym==0.21.0
numpy==1.21.5
Pillow==8.2.0
stable_baselines3==1.6.2
torch==1.12.1
torchvision==0.13.1
